os/board/rtl8730e: Sync to 8.3c SDK for TizenRT WiFi specific: -Bug fixes for WiFi stability -Enable WiFi high throughput -Adjust WiFi power by rate for Samsung Others: -Synchronise code for Peripheral, BLE, WiFi based on 8.3c SDK
SDK 8.3c : major update
Added -Werror option for the AP makefile
Upgrade lwip to version 2.1.2
Upgrade the audio feature
Upgrade and Fix bluetooth related feature
Replace the fatfs and littlefs example with the vfs example
Optimize wifi-audio layer
Modify OTA HTTP tool and Example
Simplify elfbin/imgtool.sh and unify rdpiv and rsip iv
Remove the media cache manager
Update pre-build lib
some WiFi related bug fix, new structure & ipc layer changes

+++ os/board/rtl8730e/src/component/mbed/hal/analogin_api.h
@@ -82,14 +82,14 @@ uint16_t analogin_read_u16(analogin_t *obj);
   * @param adc_data: ADC conversion data from normal channel.
   * @return Normal channel voltage in mV.
   */
-uint32_t analogin_voltage_norm(uint16_t adc_data);
+uint32_t analogin_voltage_norm(uint32_t adc_data);
 
 /**
   * @brief Get channel voltage in mV according to conversion data from VBAT channel(6).
   * @param adc_data: ADC conversion data from VBAT channel.
   * @return VBAT channel voltage in mV.
   */
-uint32_t analogin_voltage_vbat(uint16_t adc_data);
+uint32_t analogin_voltage_vbat(uint32_t adc_data);
 
 ///@}
 /*\@}*/
